+++ CHANGELOG.md
@@ -13,6 +13,7 @@ BUG FIXES:
       fail [GH-3391]
   - core: `vagrant` on Windows cmd.exe doesn't always exit with exit
       code zero. [GH-3420]
+  - commands/package: Nice error if includes contain symlinks. [GH-3200]
   - commands/rsync-auto: Don't crash if the machine can't be communicated
       to. [GH-3419]
   - guests/fedora: Fix hostname setting. [GH-3382]